update dirs structure + use latest js + update package.json
