+++ packages/registrar_client/package.json
@@ -5,13 +5,13 @@
   "keywords": [],
   "author": "Salma Elsoly <salmaelsoly@gmail.com>",
   "license": "ISC",
-  "main": "lib/registrar_client.js",
+  "main": "./dist/src/index.js",
   "directories": {
     "lib": "src",
     "test": "tests"
   },
   "files": [
-    "lib"
+    "dist"
   ],
   "repository": {
     "type": "git",
@@ -20,7 +20,8 @@
   "scripts": {
     "test": "jest",
     "format": "prettier --write .",
-    "lint": "eslint src/**/*.ts"
+    "lint": "eslint src/**/*.ts",
+    "build": "tsc"
   },
   "bugs": {
     "url": "https://github.com/threefoldtech/tfgridv4-sdk-ts/issues"
