+++ packages/registrar_client/src/modules/zos.ts
@@ -1,4 +1,4 @@
-import { RegistrarClient } from "../../client/client";
+import { RegistrarClient } from "../client/client";
 
 export class Zos {
   private client: RegistrarClient;
@@ -13,9 +13,8 @@ export class Zos {
     try {
       const data = await this.client.get<string>(`${this.zosUri}/version`);
       return data;
-    } catch (e) {
-      console.error("Failed to get zos version: ", e);
-      return null;
+    } catch (e: any) {
+      throw new Error(`Failed to get zos version: ${e.response?.status} ${e.response?.statusText}`);
     }
   }
 }
